+++ examples/multiple-target-groups/main.tf
@@ -1,3 +1,11 @@
+terraform {
+  required_version = ">= 1.0.11"
+
+  required_providers {
+    aws = ">= 4.0.0"
+  }
+}
+
 provider "aws" {
   region = "eu-west-1"
 }
@@ -136,7 +144,7 @@ module "fargate" {
   target_groups = [
     {
       target_group_name = "external-alb"
-      container_port    = 80
+      container_port    = 443
     },
     {
       target_group_name = "internal-alb"
@@ -186,4 +194,3 @@ resource "aws_security_group_rule" "test_sg_ingress" {
   to_port                  = 3022
   source_security_group_id = module.fargate.service_sg_id
 }
-